+++ src/main/java/com/kruemblegard/block/FalseWaystoneBlock.java
@@ -1,7 +1,0 @@
-package com.kruemblegard.block;
-
-public class FalseWaystoneBlock extends HauntedWaystoneBlock {
-    public FalseWaystoneBlock(Properties properties) {
-        super(properties);
-    }
-}